requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-jazzmin==2.6.0
geographiclib==2.0
geopy==2.4.1
idna==3.6
lxml==5.1.0
pillow==10.2.0
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.15.2
python-dotenv==1.0.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.1.0
